Rename P2P peer_directory folder to peer_book
